[+] moved test2GetTextLinks
